requirements.txt
streamlit==1.27.0
networkx==3.1
matplotlib==3.7.2
pandas==2.0.3
Pillow==10.0.0
reportlab==4.0.4